[qa] Refactor tests needing credentials into individual modules
